Remove all whitespace instead of just strip
urlsafe_encode64 incorrectly encodes credentials that occasionally
encode64 to have a `/` in them.  Instead of using a `/`, it replaces it
with a `_`.

The behavior exhibited with the credentials occasionally not working
with encode64 was likely a newline appearing in the middle of the
output.  urlsafe_encode64 also removes newlines from the output. Since
sumologic isn't expecting any whitespace in the base64 encoded output, we should just
gsub it away.